[ADD] change-user-pwd, restore, deploy-dev.yml
Redesigned the way user password is changed. Now it can be changed via
shell, so no old password is needed. It is also now separate playbook,
which can be used to change password for any user on any environment.

Restore playbook allows to restore database for specific environment.
deploy-dev playbook allows to deploy to development environment. It is
different from other deploy playbooks that it does not do any
backups/restores, but drops database before updating, to make sure that
we always have new data.

Also refactored the way roles were organized, to better reuse it.