[FIX] certbot renew
Standard `certbot renew` does not work, it always fails with error
saying challenge failed (might be related that everything is set up
using containers).

So now we use different approach and rely on `certonly` command that
was previously used just to create new certificates. But this command
also renew when need to. So we now have same script run when deploying
and when running cron for renew.