[FIX] delegated_restore: re-set web.base.url
When we access odoo shell with different port, to reset params, it will
use that port for re-set. And we don't want custom port to be used for
that. So we re-reset to default port.

This is limitation of our odoo docker image, that we can't properly
access odoo shell when main odoo instance is running with same port (
with docker-compose run). Normal fix would be to redesign the way we
can use commands within odoo docker image.