[FIX] pre_init, container: odoo init
Odoo db in some cases would not fully init and would crash because it
would start using it too early. This was causes when init was started
in background without waiting for it to finish, so other tasks would
run before db was fully created.

Now we wait to make sure init is complete.